Move get_mpi to utils
